move Page component form components directory to layouts
