+++ src/frontend/web_application/src/layouts/Page/components/CallToAction/index.spec.jsx
