+++ src/frontend/web_application/src/layouts/Page/components/Header/components/SearchField/index.jsx
