+++ src/frontend/web_application/src/layouts/Page/components/Header/components/SearchField/style.scss
