+++ src/frontend/web_application/src/layouts/Page/components/Header/style.scss
