+++ src/frontend/web_application/src/layouts/Page/components/MainView/style.scss
