+++ src/frontend/web_application/src/layouts/Page/components/Navigation/components/ApplicationSwitcher/presenter.jsx
