+++ src/frontend/web_application/src/layouts/Page/components/Navigation/components/Navbar/components/Link/style.scss
