+++ src/frontend/web_application/src/layouts/Page/components/Navigation/components/Navbar/components/Tab/style.scss
