+++ src/frontend/web_application/src/layouts/Page/components/Navigation/components/Navbar/index.jsx
