+++ src/frontend/web_application/src/layouts/Page/components/Navigation/style.scss
