+++ src/frontend/web_application/src/layouts/Page/components/NavigationAlt/presenter.jsx
