+++ src/frontend/web_application/src/layouts/Page/components/NavigationAlt/style.scss
