+++ src/frontend/web_application/src/layouts/Page/components/NotificationCenter/style.scss
