+++ src/frontend/web_application/src/layouts/Page/components/OffCanvas/index.jsx
