+++ src/frontend/web_application/src/layouts/Page/presenter.jsx
