+++ src/frontend/web_application/src/layouts/Page/style.scss
